requirements.txt
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.3.0
bleach==3.2.1
category-encoders==2.2.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.3
docutils==0.16
embedding-as-service-client==1.0.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
greenlet==1.0.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
keyring==21.4.0
lightgbm==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
patsy==0.5.1
pickle-mixin==1.0.2
pkginfo==1.5.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
Pygments==2.7.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytest-flask==1.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
pywin32-ctypes==0.2.0
pyzmq==20.0.0
readme-renderer==26.0
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.20
statsmodels==0.12.2
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.50.1
twine==3.2.0
urllib3==1.24.3
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
zmq==0.0.0
